move function and add doc
